refactor stats tests. create statistics module and move content to the module.
